Refactor campus, standing out of profile package
